+++ spec/spec_helper.rb
@@ -4,3 +4,7 @@ require 'coveralls'
 Coveralls.wear!
 
 require 'stoplight'
+
+Dir.glob(File.join('.', 'spec', 'support', '**', '*.rb')).each do |filename|
+  require filename
+end